+++ config/services_test.yaml
@@ -8,4 +8,4 @@ services:
     #
     # test.App\Service\MyService: '@App\Service\MyService'
     JsonToEntityTransformer: '@PiaApi\DataExchange\Transformer\JsonToEntityTransformer'
-    PiaTemplateService: '@PiaApi\Services\PiaTemplateService'
+    PiaTemplateService: '@PiaApi\Service\PiaTemplateService'
