+++ src/Command/CreateApplicationCommand.php
@@ -10,7 +10,7 @@
 
 namespace PiaApi\Command;
 
-use PiaApi\Services\ApplicationService;
+use PiaApi\Service\ApplicationService;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
