+++ src/Controller/BackOffice/PiaTemplateController.php
@@ -14,7 +14,7 @@ use PiaApi\Entity\Pia\PiaTemplate;
 use PiaApi\Form\PiaTemplate\CreatePiaTemplateForm;
 use PiaApi\Form\PiaTemplate\EditPiaTemplateForm;
 use PiaApi\Form\PiaTemplate\RemovePiaTemplateForm;
-use PiaApi\Services\PiaTemplateService;
+use PiaApi\Service\PiaTemplateService;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Symfony\Component\HttpFoundation\Request;
