+++ src/Controller/BackOffice/StructureController.php
@@ -20,8 +20,8 @@ use PiaApi\Form\Structure\EditStructureTypeForm;
 use PiaApi\Form\Structure\RemoveStructureForm;
 use PiaApi\Form\Structure\RemoveStructureTypeForm;
 use PiaApi\Form\User\CreateUserForm;
-use PiaApi\Services\StructureService;
-use PiaApi\Services\StructureTypeService;
+use PiaApi\Service\StructureService;
+use PiaApi\Service\StructureTypeService;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Symfony\Component\HttpFoundation\Request;
