+++ src/Service/PiaTemplateService.php
@@ -8,7 +8,7 @@
  * file that was distributed with this source code.
  */
 
-namespace PiaApi\Services;
+namespace PiaApi\Service;
 
 use Symfony\Bridge\Doctrine\RegistryInterface;
 use PiaApi\Entity\Pia\PiaTemplate;
